feat: let constructors use hydra
