+++ rul_adapt/__init__.py
@@ -3,3 +3,4 @@ __version__ = "0.0.0"
 import rul_adapt.approach
 import rul_adapt.loss
 import rul_adapt.model
+import rul_adapt.construct
