+++ rul_adapt/construct/__init__.py
@@ -1,1 +1,5 @@
-from .lstm_dann import get_lstm_dann, get_lstm_dann_config, lstm_dann_from_config
+from .lstm_dann.functional import (
+    get_lstm_dann,
+    get_lstm_dann_config,
+    lstm_dann_from_config,
+)
