+++ rul_adapt/construct/lstm_dann/config/base.yaml
@@ -1,0 +1,48 @@
+dm:
+  source:
+    _target_: rul_datasets.CmapssReader
+    fd: ???
+    feature_select: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23]
+  target:
+    fd: ???
+    percent_broken: 0.8
+  batch_size: ???
+
+feature_extractor:
+  _target_: rul_adapt.model.LstmExtractor
+  input_channels: 24
+  lstm_units: ???
+  fc_units: ???
+  lstm_dropout: ???
+  fc_dropout: ???
+
+regressor:
+  _target_: rul_adapt.model.FullyConnectedHead
+  input_channels: ${feature_extractor.fc_units}
+  act_func_on_last_layer: False
+  units: ???
+  dropout: ???
+
+domain_disc:
+  _target_: rul_adapt.model.FullyConnectedHead
+  input_channels: ${feature_extractor.fc_units}
+  act_func_on_last_layer: False
+  units: ???
+  dropout: ???
+
+dann:
+  _target_: rul_adapt.approach.DannApproach
+  lr_decay_factor: 0.1
+  lr_decay_epochs: 100
+  dann_factor: ???
+  lr: ???
+  weight_decay: ???
+
+trainer:
+  _target_: pytorch_lightning.Trainer
+  max_epochs: 200
+  gradient_clip_val: 1.0
+  callbacks:
+    - _target_: pytorch_lightning.callbacks.EarlyStopping
+      monitor: "val/target_rmse/dataloader_idx_1"
+      patience: 20
