+++ rul_adapt/construct/lstm_dann/config/task/2-1.yaml
@@ -1,0 +1,27 @@
+# @package _global_
+
+dann:
+  dann_factor: 1.0
+  lr: 0.01
+  weight_decay: 0.01
+
+dm:
+  batch_size: 512
+  source:
+    fd: 2
+  target:
+    fd: 1
+
+domain_disc:
+  dropout: 0.1
+  units: [16, 16, 1]
+
+feature_extractor:
+  fc_dropout: 0.1
+  fc_units: 64
+  lstm_dropout: 0.1
+  lstm_units: [64]
+
+regressor:
+  dropout: 0.0
+  units: [32, 1]
