+++ rul_adapt/construct/lstm_dann/config/task/3-2.yaml
@@ -1,0 +1,27 @@
+# @package _global_
+
+dann:
+  dann_factor: 2.0
+  lr: 0.01
+  weight_decay: 0.01
+
+dm:
+  batch_size: 256
+  source:
+    fd: 3
+  target:
+    fd: 2
+
+domain_disc:
+  dropout: 0.1
+  units: [32, 32, 1]
+
+feature_extractor:
+  fc_dropout: 0.3
+  fc_units: 64
+  lstm_dropout: 0.3
+  lstm_units: [64, 32]
+
+regressor:
+  dropout: 0.1
+  units: [32, 32, 1]
